feat: restyle delete account page
