+++ lib/core/routes.dart
@@ -15,7 +15,7 @@ import 'package:appdonationsgestor/pages/forgot_password_page.dart';
 import 'package:appdonationsgestor/pages/register_pages/finalize_registration_page.dart';
 import 'package:appdonationsgestor/pages/favorites_page.dart';
 import 'package:appdonationsgestor/pages/profile_pages/manager_profile_page.dart';
-import 'package:appdonationsgestor/pages/settings_pages/remove_account_page.dart';
+import 'package:appdonationsgestor/pages/settings_pages/delete_account_pages/delete_account_page.dart';
 import 'package:appdonationsgestor/pages/settings_pages/edit_profile_page.dart';
 import 'package:appdonationsgestor/pages/settings_pages/notifications_page.dart';
 import 'package:appdonationsgestor/pages/necessity_page.dart';
@@ -45,7 +45,7 @@ class RouteNames {
   static const String finalizeRegistrationPage = "finalizeRegistrationPage";
   static const String favoritesPage = "favoritesPage";
   static const String managerProfilePage = "managerProfilePage";
-  static const String removeAccountPage = "removeAccountPage";
+  static const String deleteAccountPage = "deleteAccountPage";
   static const String editProfilePage = "editProfilePage";
   static const String necessityPage = "necessityPage";
   static const String hystoryPage = "hystoryPage";
@@ -194,10 +194,10 @@ class AppRountersConfiguration {
               );
             }),
         GoRoute(
-          path: '/removeAccountPage',
-          name: RouteNames.removeAccountPage,
+          path: '/deleteAccountPage',
+          name: RouteNames.deleteAccountPage,
           pageBuilder: (context, state) => const MaterialPage(
-            child: RemoveAccountPage(),
+            child: DeleteAccountPage(),
           ),
         ),
         GoRoute(
