--- lib/pages/settings_pages/remove_account_page.dart
@@ -1,169 +1,0 @@
-import 'package:appdonationsgestor/auth/app_data.dart';
-import 'package:appdonationsgestor/auth/auth_service.dart';
-import 'package:appdonationsgestor/components/custom_text_field.dart';
-import 'package:appdonationsgestor/resources/constant_colors.dart';
-import 'package:appdonationsgestor/resources/text_styles.dart';
-import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
-
-class RemoveAccountPage extends StatefulWidget {
-  const RemoveAccountPage({super.key});
-
-  @override
-  State<RemoveAccountPage> createState() => _RemoveAccountPage();
-}
-
-class _RemoveAccountPage extends State<RemoveAccountPage> {
-  final TextEditingController emailController = TextEditingController();
-  final TextEditingController passwordController = TextEditingController();
-
-  @override
-  void dispose() {
-    emailController.dispose();
-    passwordController.dispose();
-    super.dispose();
-  }
-
-  void deleteAccount() async {
-    try {
-      await authService.value.deleteAccount(
-          email: emailController.text, password: passwordController.text);
-      AppData.navBarCurrentIndexNotifier.value = 0;
-      AppData.onboardingCurrentIndexNotifier.value = 0;
-      if (context.mounted) {
-        context.go('/');
-      }
-    } catch (e) {
-      print(e.toString());
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          leading: IconButton(
-            icon: const Icon(
-              Icons.arrow_back_ios,
-              color: ConstantsColors.blueShade900,
-            ),
-            onPressed: () => Navigator.pop(context),
-          ),
-          title: Text(
-            'Excluir conta',
-            style: const TextStyle(
-              color: ConstantsColors.blueShade900,
-              fontSize: 24,
-            ).merge(TextStylesConstants.kinterSemiBold),
-          ),
-          centerTitle: true,
-        ),
-        body: Stack(
-          children: [
-            SizedBox.expand(
-              child: Container(
-                decoration: const BoxDecoration(
-                  gradient: LinearGradient(
-                    colors: [
-                      ConstantsColors.blueShade500,
-                      ConstantsColors.tealShade200,
-                    ],
-                    begin: Alignment.bottomCenter,
-                    end: Alignment.topCenter,
-                  ),
-                ),
-                child: Padding(
-                  padding: const EdgeInsets.all(12.0),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      const SizedBox(height: 40),
-                      Text(
-                        "Deletar minha conta",
-                        style: TextStylesConstants.kpoppinsBold.merge(
-                            const TextStyle(
-                                fontSize: 32.0,
-                                color: ConstantsColors.whiteShade900)),
-                      ),
-                      const SizedBox(height: 100),
-                      Column(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          CustomTextFields(
-                            icon: Icons.email,
-                            label: 'Digite seu email',
-                            secret: false,
-                            controller: emailController,
-                            keyboardType: TextInputType.emailAddress,
-                          ),
-                          const SizedBox(height: 20),
-                          Stack(
-                            children: [
-                              CustomTextFields(
-                                icon: Icons.lock,
-                                label: 'Digite sua senha',
-                                secret: true,
-                                controller: passwordController,
-                                keyboardType: TextInputType.visiblePassword,
-                              ),
-                            ],
-                          ),
-                          const SizedBox(height: 20),
-                          Column(
-                            children: [
-                              Padding(
-                                padding:
-                                    const EdgeInsets.symmetric(horizontal: 8.0),
-                                child: ElevatedButton(
-                                  style: ElevatedButton.styleFrom(
-                                    backgroundColor:
-                                        ConstantsColors.redShade900,
-                                    shape: RoundedRectangleBorder(
-                                      borderRadius: BorderRadius.circular(15),
-                                    ),
-                                  ),
-                                  onPressed: () {
-                                    deleteAccount();
-                                  },
-                                  child: Text(
-                                    'Deletar permanentemente',
-                                    style: const TextStyle(
-                                      color: ConstantsColors.whiteShade900,
-                                      fontSize: 15,
-                                    ).merge(TextStylesConstants.kpoppinsMedium),
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      )
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            Positioned(
-              top: 0,
-              left: 0,
-              right: 0,
-              child: SafeArea(
-                child: Padding(
-                  padding: const EdgeInsets.only(left: 8.0, top: 8.0),
-                  child: Align(
-                    alignment: Alignment.topLeft,
-                    child: IconButton(
-                      icon:
-                          const Icon(Icons.arrow_back_ios, color: Colors.white),
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ));
-  }
-}
